requirements.txt
torch>=1.9.0
torchvision>=0.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
pandas>=1.3.0
scikit-learn>=1.0.0
python-multipart>=0.0.6
requests>=2.28.0
beautifulsoup4>=4.11.0
google-generativeai>=0.3.0
python-multipart>=0.0.6